# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Intelligent Retail AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: minPrice
          in: query
          required: false
          schema:
            type: number
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
        - name: inStock
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
        "404":
          description: Product not found

  # Categories
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerBody"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Customer not found

  # Orders
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: customerId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found

  # Inventory
...
  /inventory/{productId}:
    patch:
      operationId: updateInventory
      tags: [inventory]
      summary: Update inventory quantity for a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInventoryBody"
      responses:
        "200":
          description: Inventory updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"

  # AI Features
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        sku:
          type: string
        imageUrl:
          type: ["string", "null"]
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isActive:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - price
        - sku
        - stock
        - lowStockThreshold
        - isActive
        - reviewCount
        - createdAt
        - updatedAt
    CreateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        sku:
          type: string
        imageUrl:
          type: ["string", "null"]
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isActive:
          type: boolean
      required:
        - name
        - price
        - sku
        - stock
        - lowStockThreshold
        - isActive
    UpdateProductBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        sku:
          type: string
        imageUrl:
          type: ["string", "null"]
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        isActive:
          type: boolean
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        totalOrders:
          type: integer
        totalSpent:
          type: number
        lastOrderAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - totalOrders
        - totalSpent
        - createdAt
    CreateCustomerBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
      required:
        - name
        - email
    Order:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
        total:
          type: number
        itemCount:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - status
        - total
        - itemCount
        - items
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        price:
          type: number
      required:
        - id
        - productId
        - productName
        - quantity
        - price
    CreateOrderBody:
      type: object
      properties:
        customerId:
          type: ["integer", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/CreateOrderItemBody"
        notes:
          type: ["string", "null"]
      required:
        - items
    CreateOrderItemBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required:
        - productId
        - quantity
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, processing, shipped, delivered, cancelled]
      required:
        - status
    InventoryItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        sku:
          type: string
        currentStock:
          type: integer
        lowStockThreshold:
          type: integer
        status:
          type: string
          enum: [healthy, low, critical, overstock]
        lastRestockedAt:
          type: ["string", "null"]
      required:
        - productId
        - productName
        - sku
        - currentStock
        - lowStockThreshold
        - status
...
    UpdateInventoryBody:
      type: object
      properties:
        quantity:
          type: integer
        operation:
          type: string
          enum: [set, add, subtract]
      required:
        - quantity
        - operation
...