# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intelligent Retail AI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderItemBody } from "./createOrderItemBody";

export interface CreateOrderBody {
  /** @nullable */
  customerId?: number | null;
  items: CreateOrderItemBody[];
  /** @nullable */
  notes?: string | null;
}