# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Intelligent Retail AI Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from "./orderItem";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: number;
  /** @nullable */
  customerId?: number | null;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  itemCount: number;
  items: OrderItem[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}